1.增加超级密码 2.开发时间6秒 3.有效日期的计算 4.阀门生产日期 5.初始化ID值，生产日期，有效日期

+++ slave/Bsp/delay.h
@@ -6,13 +6,13 @@
 #define BIT(x)	(1 << (x))
 #endif
 
-void delay_ms(uint16_t nms);
+void delay_ms(uint32_t nms);
 void delay_us(uint32_t nus);
 void Delay(uint32_t count);
-void Delay_Init(uint8_t SYSCLK);
+void Delay_Init(uint32_t SYSCLK);
 
-extern uint16_t fac_ms;//全局变量
-extern uint8_t fac_us;//全局变量
+extern uint32_t fac_ms;//全局变量
+extern uint32_t fac_us;//全局变量
 
 
 #endif
